pyproject.toml
[project]
name = "satellite-segmentation"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geojson>=3.2.0",
    "h5py>=3.15.1",
    "httpx>=0.28.1",
    "lightning>=2.6.1",
    "omegaconf>=2.3.0",
    "osmium>=4.3.0",
    "pillow>=12.1.1",
    "pyproj>=3.7.2",
    "rootutils>=1.0.7",
    "shapely>=2.1.2",
    "supermercado>=0.3.0",
    "flake8-pyproject>=1.2.4",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
    "wemake-python-styleguide>=1.5.0",
    "mercantile>=1.2.1",
    "dvc[ssh]>=3.66.1",
    "torchvision",
]

[tool.ruff]
# Включаем стандартные правила (Pyflakes, pycodestyle)
select = ["E", "F", "W", "C90", "I"]
# Игнорируем правила, которые будут перекрываться проверками WPS
ignore = ["E402"]
target-version = "py311"
line-length = 120
indent-width = 4
output-format = "full"
show-fixes = true
exclude = [
    "scripts/*.py",
]

[tool.flake8]
exclude = [
    "scripts/*.py",
    "src/*py",
]
# Оставляем только правила WPS, так как остальное проверит Ruff
select = "WPS"
# Настройка специфичных для WPS параметров
max-line-length = 120
show-source = true
ignore = ["WPS111", "WPS420", "WPS100"]
max-methods = 10
max-line-complexity = 20
max-module-members = 10

[tool.isort]
profile = "wemake"
line_length = 120
src_paths = ["src"]

[tool.mypy]
exclude = [
    "^.venv/",
    "^scripts/",
    "^src/",
]
explicit_package_bases = true
